refactor: MyFont 컴포넌트 리팩토링 (#43)
CustomFontList를 widgets/font-list 로 이동
FontRankingChart를 widgets/chart로 이동
url을 my-fonts에서 my-font로 수정
