+++ src/shared/ui/SideNavigationBar.jsx
@@ -17,7 +17,7 @@ const NAV_LIST = [
   { label: '폰트 합성', Icon: NavFontSynthesisIcon, url: '/font-synthesis' },
   { label: '둘러보기', Icon: NavExploreIcon, url: '/explore' },
   { label: '북마크', Icon: NavBookmarkIcon, url: '/bookmark' },
-  { label: '내 폰트', Icon: NavMyFontIcon, url: '/my-fonts' },
+  { label: '내 폰트', Icon: NavMyFontIcon, url: '/my-font' },
 ]
 
 export const SideNavigationBar = () => {
