+++ src/widgets/chart/ui/FontRankingChart.jsx
@@ -1,8 +1,9 @@
 import styled from 'styled-components'
 
 import { EmptyMessage } from '@/shared/ui/EmptyMessage'
-import { BarChart } from '@/widgets/chart/ui/BarChart'
-import { ChartLegend } from '@/widgets/chart/ui/ChartLegend'
+
+import { BarChart } from './BarChart'
+import { ChartLegend } from './ChartLegend'
 
 const dummyRank = [
   { name: '가나다체', bookmark: 33, download: 19 },
@@ -15,7 +16,7 @@ const dummyRank = [
 export const FontRankingChart = () => {
   const data = dummyRank
 
-  if (!data || !data.length) return <EmptyMessage message="등록된 폰트가 없습니다." />
+  if (!data || !data.length) return <EmptyMessage message="아직 등록된 폰트가 없습니다." />
 
   return (
     <Container>
